Files in movies DTOs moved.
